verify should only check ttl if > 0
The docs state the TTL is only checked during verify if > 0. This is similar to how Python does it too (well, Python checks if ttl is not None). I was trying to figure out what was wrong with my code, then realized it was because of this)!

If you think this should stay >= 0, then I am happy to update the docs!